more doubles

+++ ParallelPageRank/src/getRank.h
@@ -13,8 +13,8 @@
 void makeP(double *Avals, MKL_INT *rowind, MKL_INT *numRow, MKL_INT *colind, MKL_INT *nnz,  double dP);
 void getRank(double *Pvals, double *x, MKL_INT *rowind, MKL_INT *colind, MKL_INT *numRows, MKL_INT *nnz, double tol, float dP);
 //void makeSinks(MKL_INT *rowind, MKL_INT *colind, float *d, MKL_INT numRow);
-float sum(double *x, int N);
+double sum(double *x, int N);
 void ones(double *a, int N);
-float getError(double *v1, double *v2, MKL_INT size);
+double getError(double *v1, double *v2, MKL_INT size);
 
 #endif /* getRank_h */
